+++ lib/aixm/refinements.rb
@@ -15,6 +15,21 @@ module AIXM
       "Ø" => "Oe"
     }.freeze
 
+    # @!method then_if
+    #   Same as +Object#then+ but only applied if the condition is true.
+    #
+    #   @example
+    #     "foobar".then_if(false) { |s| s.gsub(/o/, 'i') }   # => "foobar"
+    #     "foobar".then_if(true) { |s| s.gsub(/o/, 'i') }    # => "fiibar"
+    #
+    #   @note This is a refinement for +Object+
+    #   @return [Object]
+    refine Object do
+      def then_if(condition, &block)
+        condition ? self.then(&block) : self
+      end
+    end
+
     # @!method to_digest
     #   Builds a 4 byte hex digest from the Array payload.
     #